Develop

See merge request s14-final/S14P31S309!297

+++ frontend/app/src/main/java/com/ssafy/s309/data/repository/source/SamsungHealthDataSource.kt
@@ -36,7 +36,7 @@ class SamsungHealthDataSource
         override suspend fun getTodaySummary(): DailyHealthSummary? {
             val mgr = holder.manager ?: return null
             return try {
-                val steps = mgr.getTodaySteps().toInt()
+                val steps = mgr.getTodaySteps()
                 val sleep = mgr.getLastSleepDurationMinutes()
                 if (steps == 0 && sleep == 0) {
                     null
